Added Gaussian input creation from template
Now with the ‘-x template’ flag and an appropriate template file, the
program can create Gaussian input files, then submit them. Works for
batches as well.
Also updated the way it removes file extensions from file names to
hopefully be more general (should now allow other ‘.’s in the file
name). Finally (I think), if it finds multiple files based on the input
name, but it wasn’t input as a batch job, it will ask if you want to do
it as a batch job and just forgot the ‘-b’ flag.