Support for (TD)DFT from Gaussian output
Only set to explicitly recognize the PBE1PBE (PBE0) functional so far,
but any other Gaussian output read will default to being processed as
if it is DFT. Note, it is necessary to declare the method before a
slash as in “B3LYP/6-31G” for it to recognize the method. I don’t see
how to get around this other than making some sort of library to
recognize all methods available to Gaussian (that might be used, at
least).