take status file to restart; all args are kwargs with None as default
Also, now self.args is a dict (might be a better way to do this as
opposed to listing all the arguments at least three times here).
If not rerunning, make sure all arguments are given (and not None).
Properly use self.status to see if the job is being rerun.